Cleanup and re-try CI

+++ include/boost/multiprecision/detail/float_string_cvt.hpp
@@ -157,12 +157,12 @@ std::string convert_to_string(Backend b, std::streamsize digits, std::ios_base::
             // Bankers rounding:
             if ((*result.rbegin() - '0') & 1)
             {
-               round_string_up_at(result, result.size() - 1, expon);
+               round_string_up_at(result, (int) result.size() - 1, expon);
             }
          }
          else if (cdigit >= 5)
          {
-            round_string_up_at(result, result.size() - 1, expon);
+            round_string_up_at(result, (int) result.size() - 1, expon);
          }
       }
    }
